remove second client
