+++ stpmex/__init__.py
@@ -1,4 +1,4 @@
-__all__ = ['__version__', 'Client', 'ClientEfws']
+__all__ = ['__version__', 'Client']
 
-from .client import Client, ClientEfws
+from .client import Client
 from .version import __version__
